+++ electron.vite.config.ts
@@ -1,0 +1,21 @@
+import path from "node:path";
+import { defineConfig, externalizeDepsPlugin } from "electron-vite";
+import vue from "@vitejs/plugin-vue";
+
+export default defineConfig({
+  main: {
+    plugins: [externalizeDepsPlugin()]
+  },
+  preload: {
+    plugins: [externalizeDepsPlugin()]
+  },
+  renderer: {
+    plugins: [vue()],
+    resolve: {
+      alias: {
+        "@renderer": path.resolve("src/renderer/src"),
+        "@shared": path.resolve("src/shared")
+      }
+    }
+  }
+});